requirements.txt
Flask==2.2.5
altair==4.2.2
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.7.0
anyio==4.6.2
appdirs==1.4.4
beautifulsoup4==4.12.3
bs4==0.0.2
fake-useragent==1.5.1
frozenlist==1.4.1
fsspec==2024.9.0
GitPython==3.1.31
greenlet==3.0.3
h11==0.14.0
html2text==2024.2.26
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.4
itsdangerous==2.2.0
Jinja2==3.1.2
litellm==1.48.0
lxml==5.3.0
markdown-it-py==3.0.0
multidict==6.1.0
numpy==1.26.4
openai==1.51.2
pandas==2.0.3
playwright==1.47.0
pyee==12.0.0
Pygments==2.15.1
PyJWT==2.9.0
pypdf==5.0.1
pyquery==2.0.1
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.32.2
requests-html==0.10.0
selenium==4.25.0
streamlit==1.24.1
tiktoken==0.8.0
tqdm==4.66.5
twilio==9.3.3
urllib3==1.26.20
validators==0.20.0
watchdog==3.0.1
websocket-client==1.8.0
Werkzeug==3.0.4

